refactor: tighten codex provider boundaries
