+++ packages/integrations/anyagent/package.json
@@ -19,6 +19,7 @@
     "zod": "^4.3.6"
   },
   "devDependencies": {
+    "@types/node": "^22.0.0",
     "typescript": "^5.8.0",
     "vitest": "^4.1.2"
   }
